finish migrations

+++ cli/src/main/kotlin/dev/whyoleg/ktd/cli/api/ApiCommand.kt
@@ -18,7 +18,7 @@ class ApiCommand : Subcommand("api") {
         println("Files downloaded")
         val data = parseData(tdBytes, tlBytes)
         data.forEach { (kind, scheme) ->
-            if (kind == TlKind.Ignore) return@forEach
+            if (kind == TlKind.Ignore || kind == TlKind.All) return@forEach
             File("api/ktd-api-${kind.name.toLowerCase()}/src/commonMain/kotlin").deleteRecursively()
             File("api-suspend/ktd-api-${kind.name.toLowerCase()}-suspend/src/commonMain/kotlin").deleteRecursively()
 
@@ -27,7 +27,10 @@ class ApiCommand : Subcommand("api") {
             builderFile(scheme, kind)
             suspendTdApiFunctions(scheme, kind)
         }
-        oldApiFile(data.getValue(TlKind.Core).copy(data = data.flatMap { it.value.data }))
+        File("migration/v060/ktd-api-raw/src/commonMain/kotlin").deleteRecursively()
+        File("migration/v060/ktd-api-coroutines/src/commonMain/kotlin").deleteRecursively()
+        oldApiFile(data.getValue(TlKind.All))
+        oldSuspendTdApiFunctions(data.getValue(TlKind.All))
         println("New api saved")
     }
